Removed CheckOverflow override from the common GpuOverrides and moving it o the shimmed one because of CheckOverflow changes in 340
