+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/GpuOverrides.scala
@@ -896,72 +896,7 @@ object GpuOverrides extends Logging {
           }
         }
       }),
-    expr[CheckOverflow](
-      "CheckOverflow after arithmetic operations between DecimalType data",
-      ExprChecks.unaryProjectInputMatchesOutput(TypeSig.DECIMAL_128,
-        TypeSig.DECIMAL_128),
-      (a, conf, p, r) => new ExprMeta[CheckOverflow](a, conf, p, r) {
-        private[this] def extractOrigParam(expr: BaseExprMeta[_]): BaseExprMeta[_] =
-          expr.wrapped match {
-            case lit: Literal if lit.dataType.isInstanceOf[DecimalType] =>
-              // Lets figure out if we can make the Literal value smaller
-              val (newType, value) = lit.value match {
-                case null =>
-                  (DecimalType(0, 0), null)
-                case dec: Decimal =>
-                  val stripped = Decimal(dec.toJavaBigDecimal.stripTrailingZeros())
-                  val p = stripped.precision
-                  val s = stripped.scale
-                  val t = if (s < 0 && !SQLConf.get.allowNegativeScaleOfDecimalEnabled) {
-                    // need to adjust to avoid errors about negative scale
-                    DecimalType(p - s, 0)
-                  } else {
-                    DecimalType(p, s)
-                  }
-                  (t, stripped)
-                case other =>
-                  throw new IllegalArgumentException(s"Unexpected decimal literal value $other")
-              }
-              expr.asInstanceOf[LiteralExprMeta].withNewLiteral(Literal(value, newType))
-            case _ => expr
-          }
-        private[this] lazy val binExpr = childExprs.head
-        private[this] lazy val lhs = extractOrigParam(binExpr.childExprs.head)
-        private[this] lazy val rhs = extractOrigParam(binExpr.childExprs(1))
-        private[this] lazy val lhsDecimalType =
-          DecimalUtil.asDecimalType(lhs.wrapped.asInstanceOf[Expression].dataType)
-        private[this] lazy val rhsDecimalType =
-          DecimalUtil.asDecimalType(rhs.wrapped.asInstanceOf[Expression].dataType)
 
-        override def convertToGpu(): GpuExpression = {
-          // Prior to Spark 3.4.0
-          // Division and Multiplication of Decimal types is a little odd. Spark will cast the
-          // inputs to a common wider value where the scale is the max of the two input scales,
-          // and the precision is max of the two input non-scale portions + the new scale. Then it
-          // will do the divide or multiply as a BigDecimal value but lie about the return type.
-          // Finally here in CheckOverflow it will reset the scale and check the precision so that
-          // Spark knows it fits in the final desired result.
-          // Here we try to strip out the extra casts, etc to get to as close to the original
-          // query as possible. This lets us then calculate what CUDF needs to get the correct
-          // answer, which in some cases is a lot smaller.
-
-          a.child match {
-            case _: Divide =>
-              // GpuDecimalDivide includes the overflow check in it.
-              GpuDecimalDivide(lhs.convertToGpu(), rhs.convertToGpu(), wrapped.dataType)
-            case _: Multiply =>
-              // GpuDecimal*Multiply includes the overflow check in it
-              val intermediatePrecision =
-                GpuDecimalMultiply.nonRoundedIntermediatePrecision(lhsDecimalType,
-                  rhsDecimalType, a.dataType)
-              GpuDecimalMultiply(lhs.convertToGpu(), rhs.convertToGpu(), wrapped.dataType,
-                useLongMultiply = intermediatePrecision > DType.DECIMAL128_MAX_PRECISION)
-            case _ =>
-              GpuCheckOverflow(childExprs.head.convertToGpu(),
-                wrapped.dataType, wrapped.nullOnOverflow)
-          }
-        }
-      }),
     expr[ToDegrees](
       "Converts radians to degrees",
       ExprChecks.mathUnary,
